cleanup: move `getColorByImage` to a util
