+++ src/utils/getColorByImage.ts
@@ -1,0 +1,21 @@
+import Color from "color"
+import ColorThief from "colorthief/dist/color-thief.mjs"
+
+const getColorByImage = (imageUrl: string) =>
+  new Promise<string>((resolve, _) => {
+    const colorThief = new ColorThief()
+
+    const imgEl = document.createElement("img")
+    imgEl.src = imageUrl
+    imgEl.width = 64
+    imgEl.height = 64
+    imgEl.crossOrigin = "anonymous"
+
+    imgEl.addEventListener("load", () => {
+      const dominantRgbColor = colorThief.getColor(imgEl)
+      const dominantHexColor = Color.rgb(dominantRgbColor).hex()
+      resolve(dominantHexColor)
+    })
+  })
+
+export default getColorByImage
